+++ pkg/execute/mockExecute.go
@@ -7,10 +7,9 @@ import (
 	"github.com/stretchr/testify/mock"
 )
 
-// MockExecute defines a simple executor for testing purposal
+// MockExecute is a mock of Execute interface
 type MockExecute struct {
 	mock.Mock
-	//	Write io.Writer
 }
 
 // NewMockExecute returns a new instance of MockExecute
@@ -18,18 +17,8 @@ func NewMockExecute() *MockExecute {
 	return &MockExecute{}
 }
 
-// Execute takes a command and args and runs it, streaming output to stdout
+// Execute is a mock
 func (e *MockExecute) Execute(ctx context.Context, command []string, resultsFunc stdoutcallback.StdoutCallbackResultsFunc, options ...ExecuteOptions) error {
 	args := e.Called(ctx, command, resultsFunc, options)
 	return args.Error(0)
-
-	// if e.Write == nil {
-	// 	e.Write = os.Stdout
-	// }
-	// if command[0] == "error" {
-	// 	return errors.New("(MockExecute::Execute) error")
-	// }
-
-	// fmt.Fprintf(e.Write, "%v", command)
-	// return nil
 }