+++ examples/ansibleplaybook-json-stdout/ansibleplaybook-json-stdout.go
@@ -1,0 +1,59 @@
+package main
+
+import (
+	"bytes"
+	"context"
+	"fmt"
+	"io"
+
+	"github.com/apenella/go-ansible/pkg/execute"
+	"github.com/apenella/go-ansible/pkg/execute/measure"
+	"github.com/apenella/go-ansible/pkg/options"
+	"github.com/apenella/go-ansible/pkg/playbook"
+	"github.com/apenella/go-ansible/pkg/stdoutcallback/results"
+)
+
+func main() {
+
+	var err error
+	var res *results.AnsiblePlaybookJSONResults
+
+	buff := new(bytes.Buffer)
+
+	ansiblePlaybookConnectionOptions := &options.AnsibleConnectionOptions{
+		Connection: "local",
+	}
+
+	ansiblePlaybookOptions := &playbook.AnsiblePlaybookOptions{
+		Inventory: "127.0.0.1,",
+	}
+
+	executorTimeMeasurement := measure.NewExecutorTimeMeasurement(
+		execute.NewDefaultExecute(
+			execute.WithWrite(io.Writer(buff)),
+		),
+	)
+
+	playbooksList := []string{"site1.yml", "site2.yml", "site3.yml"}
+	playbook := &playbook.AnsiblePlaybookCmd{
+		Playbooks:         playbooksList,
+		Exec:              executorTimeMeasurement,
+		ConnectionOptions: ansiblePlaybookConnectionOptions,
+		Options:           ansiblePlaybookOptions,
+		StdoutCallback:    "json",
+	}
+
+	err = playbook.Run(context.TODO())
+	if err != nil {
+		fmt.Println(err.Error())
+	}
+
+	res, err = results.ParseJSONResultsStream(io.Reader(buff))
+	if err != nil {
+		panic(err)
+	}
+
+	fmt.Println(res.String())
+	fmt.Println("Duration: ", executorTimeMeasurement.Duration())
+
+}