+++ examples/ansibleplaybook-with-executor-time-measurament/ansibleplaybook-with-executor-time-measurament.go
@@ -1,0 +1,53 @@
+package main
+
+import (
+	"bytes"
+	"context"
+
+	"github.com/apenella/go-ansible/pkg/execute"
+	"github.com/apenella/go-ansible/pkg/execute/measure"
+	"github.com/apenella/go-ansible/pkg/options"
+	"github.com/apenella/go-ansible/pkg/playbook"
+	"github.com/apenella/go-ansible/pkg/stdoutcallback/results"
+	"github.com/fatih/color"
+)
+
+func main() {
+
+	durationBuff := new(bytes.Buffer)
+
+	ansiblePlaybookConnectionOptions := &options.AnsibleConnectionOptions{
+		Connection: "local",
+	}
+
+	ansiblePlaybookOptions := &playbook.AnsiblePlaybookOptions{
+		Inventory: "127.0.0.1,",
+	}
+
+	executorTimeMeasurement := measure.NewExecutorTimeMeasurement(
+		execute.NewDefaultExecute(
+			execute.WithEnvVar("ANSIBLE_FORCE_COLOR", "true"),
+			execute.WithTransformers(
+				results.Prepend("Go-ansible example"),
+				results.LogFormat(results.DefaultLogFormatLayout, results.Now),
+			),
+			execute.WithShowDuration(),
+		),
+		measure.WithShowDuration(),
+		measure.WithWrite(durationBuff),
+	)
+
+	playbook := &playbook.AnsiblePlaybookCmd{
+		Playbooks:         []string{"site.yml"},
+		ConnectionOptions: ansiblePlaybookConnectionOptions,
+		Options:           ansiblePlaybookOptions,
+		Exec:              executorTimeMeasurement,
+	}
+
+	err := playbook.Run(context.TODO())
+	if err != nil {
+		panic(err)
+	}
+
+	color.Cyan("\n\t%s\n", durationBuff.String())
+}