+++ pkg/execute/interface.go
@@ -1,0 +1,12 @@
+package execute
+
+import (
+	"context"
+
+	"github.com/apenella/go-ansible/pkg/stdoutcallback"
+)
+
+// Executor interface is satisfied by those types which has a Execute(context.Context,[]string,stdoutcallback.StdoutCallbackResultsFunc,...ExecuteOptions)error method
+type Executor interface {
+	Execute(ctx context.Context, command []string, resultsFunc stdoutcallback.StdoutCallbackResultsFunc, options ...ExecuteOptions) error
+}