+++ pkg/execute/mockExecute.go
@@ -2,28 +2,23 @@ package execute
 
 import (
 	"context"
-	"errors"
-	"fmt"
-	"io"
-	"os"
 
 	"github.com/apenella/go-ansible/pkg/stdoutcallback"
+	"github.com/stretchr/testify/mock"
 )
 
-// MockExecute defines a simple executor for testing purposal
+// MockExecute is a mock of Execute interface
 type MockExecute struct {
-	Write io.Writer
+	mock.Mock
 }
 
-// Execute takes a command and args and runs it, streaming output to stdout
-func (e *MockExecute) Execute(ctx context.Context, command []string, resultsFunc stdoutcallback.StdoutCallbackResultsFunc, options ...ExecuteOptions) error {
-	if e.Write == nil {
-		e.Write = os.Stdout
-	}
-	if command[0] == "error" {
-		return errors.New("(MockExecute::Execute) error")
-	}
+// NewMockExecute returns a new instance of MockExecute
+func NewMockExecute() *MockExecute {
+	return &MockExecute{}
+}
 
-	fmt.Fprintf(e.Write, "%v", command)
-	return nil
+// Execute is a mock
+func (e *MockExecute) Execute(ctx context.Context, command []string, resultsFunc stdoutcallback.StdoutCallbackResultsFunc, options ...ExecuteOptions) error {
+	args := e.Called(ctx, command, resultsFunc, options)
+	return args.Error(0)
 }