+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
  
 ### Added
 - `ExecutorTimeMeasurement` is a decorator defined on `github.com/apenella/go-ansible/pkg/execute`, that measures the duration of an execution, it receives an `Executor` which is measured the execution time
+- Add `unreachable` state on task play results struct `AnsiblePlaybookJSONResultsPlayTaskHostsItem`
 
 ### Chanded
 - `MockExecute` uses `github.com/stretchr/testify/mock`
@@ -13,3 +14,4 @@
 
 ### Removed
 - `DefaultExecutor` does not measures the execution duration anymore. Instead of it, `ExecutorTimeMeasurement` must be used
+